requirements.txt
backports.shutil-get-terminal-size==1.0.0
colorama==0.3.7
decorator==4.0.10
Django==1.11
django-appconf==1.0.2
django-ckeditor==5.0.3
django-compressor==2.0
django-extensions==1.6.7
django-imagekit==3.3
django-ordered-model==1.2.1
django-photologue==3.5.1
django-sortedm2m==1.3.2
django-gulp==2.5.0
django-libsass

ExifRead==2.1.2
ipython==5.0.0
ipython-genutils==0.1.0
pexpect==4.2.0
pickleshare==0.7.3
pilkit==1.1.13
Pillow==3.3.0
ply==3.8
prompt-toolkit==1.0.5
ptyprocess==0.5.1
Pygments==2.1.3
pyreadline==2.1
python-slugify==1.2.0
pytz==2016.6.1
rcssmin==1.0.6
rjsmin==1.0.12
simplegeneric==0.8.1
six==1.10.0
slimit==0.8.1
traitlets==4.2.2
Unidecode==0.4.19
wcwidth==0.1.7
win-unicode-console==0.5
